requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
python-dotenv
pydantic
python-jose[cryptography]
passlib[bcrypt]
email-validator
python-multipart
alembic
pandas
openpyxl
pytest
openai
httpx
pytz
